--- Hasselman.Backsplice.Spreadsheet/Hasselman.Backsplice.Spreadsheet.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Backsplice_Spreadsheet_Interface</RootNamespace>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <PackageProjectUrl>https://github.com/james-d-hasselman/backsplice-spreadsheet-interface</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/james-d-hasselman/backsplice-spreadsheet-interface</RepositoryUrl>
-    <PackageLicenseExpression>GPL-3.0-or-later</PackageLicenseExpression>
-    <Version>$(VersionPrefix)</Version>
-    <VersionSuffix>alpha</VersionSuffix>
-    <InformationalVersion>Prerelease development package.</InformationalVersion>
-    <Authors>James D. Hasselman</Authors>
-    <AssemblyVersion></AssemblyVersion>
-  </PropertyGroup>
-
-</Project>
