+++ backsplice.spreadsheet.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "backsplice.spreadsheet", "backsplice.spreadsheet.csproj", "{15D1680B-958C-4DB0-B9E8-001251ADDC39}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{15D1680B-958C-4DB0-B9E8-001251ADDC39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15D1680B-958C-4DB0-B9E8-001251ADDC39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15D1680B-958C-4DB0-B9E8-001251ADDC39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15D1680B-958C-4DB0-B9E8-001251ADDC39}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A5808A8E-510D-47AE-99D2-2CBB4D5AC956}
+	EndGlobalSection
+EndGlobal
