+++ src/ICell.cs
@@ -1,0 +1,9 @@
+﻿// SPDX-FileCopyrightText: 2022 James D. Hasselman <james.d.hasselman@gmail.com>
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+namespace Backsplice.Spreadsheet;
+
+public interface ICell
+{
+    public string Value { get; set; }
+}
