+++ src/IColumn.cs
@@ -1,10 +1,9 @@
 ﻿// SPDX-FileCopyrightText: 2022 James D. Hasselman <james.d.hasselman@gmail.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-namespace Hasselman.Backsplice.Spreadsheet
+namespace Backsplice.Spreadsheet;
+
+public interface IColumn
 {
-    public interface ICell
-    {
-        public string Value { get; set; }
-    }
+    public double Width { get; set; }
 }
