+++ src/IRow.cs
@@ -1,10 +1,10 @@
 ﻿// SPDX-FileCopyrightText: 2022 James D. Hasselman <james.d.hasselman@gmail.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-namespace Hasselman.Backsplice.Spreadsheet
+namespace Backsplice.Spreadsheet;
+
+public interface IRow
 {
-    public interface IColumn
-    {
-        public double Width { get; set; }
-    }
+    double? Height { get; set; }
+    public IList<ICell> Cells { get; set; }
 }
